Use u8 for number of dimensions and parameters
    That should be enough.

+++ src/wavefunctions/gaussian.rs
@@ -3,16 +3,12 @@ use crate::wavefunctions::wavefunction;
 #[derive(Debug)]
 pub struct Gaussian {
     num_particles: u32,
-    num_dimensions: u32,
-    num_parameters: u32,
+    num_dimensions: u8,
+    num_parameters: u8,
 }
 
 impl wavefunction::Wavefunction for Gaussian {
-    fn new(
-        num_particles: u32,
-        num_dimensions: u32,
-        num_parameters: u32,
-    ) -> Self {
+    fn new(num_particles: u32, num_dimensions: u8, num_parameters: u8) -> Self {
         Gaussian {
             num_particles,
             num_dimensions,
